@@ -13,17 +13,17 @@
     "@radix-ui/react-dialog": "^1.1.2",
     "clsx": "^2.1.1",
     "dayjs": "^1.11.13",
-    "framer-motion": "^6.3.0",
+    "framer-motion": "^11.11.8",
     "keen-slider": "^6.8.6",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-countdown": "^2.3.6",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.0.1"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^7.0.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.11.1",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
     "@vitejs/plugin-react": "^4.3.2",
     "eslint": "^9.11.1",
     "eslint-plugin-react-hooks": "^5.1.0-rc.0",
